+++ packages/tsurlfilter/test/rules/network-rule.test.ts
@@ -1,19 +1,32 @@
 /* eslint-disable max-len */
-import { describe, it, expect } from 'vitest';
+import {
+    describe,
+    it,
+    expect,
+    vi,
+    afterAll,
+    afterEach,
+} from 'vitest';
 import {
     NetworkRuleOption,
     Request,
     RequestType,
     HTTPMethod,
     StealthOptionName,
-    setLogger,
     NetworkRuleGroupOptions,
     type NetworkRule,
 } from '../../src';
 import { createNetworkRule } from '../helpers/rule-creator';
-import { LoggerMock } from '../mocks';
 
 describe('NetworkRule constructor', () => {
+    afterEach(() => {
+        vi.clearAllMocks();
+    });
+
+    afterAll(() => {
+        vi.resetAllMocks();
+    });
+
     describe('creation of rule with $stealth modifier', () => {
         it('creates $stealth rule without options', () => {
             const rule = createNetworkRule('@@||example.org^$stealth', -1);
@@ -30,32 +43,6 @@ describe('NetworkRule constructor', () => {
             expect(stealthModifier?.hasValues()).toBeTruthy();
             expect(stealthModifier?.hasStealthOption(StealthOptionName.DoNotTrack)).toBeTruthy();
             expect(stealthModifier?.hasStealthOption(StealthOptionName.XClientData)).toBeTruthy();
-        });
-
-        it('logs debug message on $stealth rule with duplicate options', () => {
-            const loggerMock = new LoggerMock();
-            setLogger(loggerMock);
-
-            const msg = 'Duplicate $stealth modifier value "donottrack" in "donottrack|donottrack"';
-            createNetworkRule('@@||example.org^$stealth=donottrack|donottrack', 0);
-
-            expect(loggerMock.debug).toHaveBeenCalledTimes(1);
-            expect(loggerMock.debug).toHaveBeenCalledWith(msg);
-
-            setLogger(console);
-        });
-
-        it('logs debug message on $stealth modifier with no supported values', () => {
-            const loggerMock = new LoggerMock();
-            setLogger(loggerMock);
-
-            const msg = '$stealth modifier does not contain any options supported by browser extension: "webrtc|location"';
-            createNetworkRule('@@||example.org^$stealth=webrtc|location', 0);
-
-            expect(loggerMock.debug).toHaveBeenCalledTimes(1);
-            expect(loggerMock.debug).toHaveBeenCalledWith(msg);
-
-            setLogger(console);
         });
 
         it('throws error on $stealth rule with inverted options', () => {