+++ packages/tswebextension/src/lib/mv2/background/web-request-api.ts
@@ -305,9 +305,7 @@ export class WebRequestApi {
         try {
             await browser.webRequest.handlerBehaviorChanged();
         } catch (e) {
-            // TODO: use getErrorMessage instead
-            const message = e instanceof Error ? e.message : String(e);
-            logger.error(message);
+            logger.error('[tsweb.WebRequestApi.flushMemoryCache]: Cannot flush memory cache and call browser.handlerBehaviorChanged: ', e);
         }
     }
 
@@ -731,8 +729,7 @@ export class WebRequestApi {
         if (requestType === RequestType.Document || requestType === RequestType.SubDocument) {
             const frameContext = tabsApi.getFrameContext(tabId, frameId);
             if (!frameContext?.cosmeticResult) {
-                // eslint-disable-next-line max-len
-                logger.debug(`[RequestEvents.onCompleted]: cannot log script rules due to not having cosmetic result for tabId: ${tabId}, frameId: ${frameId}.`);
+                logger.debug(`[tsweb.WebRequestApi.onCompleted]: cannot log script rules due to not having cosmetic result for tabId: ${tabId}, frameId: ${frameId}.`);
                 return;
             }
 
@@ -785,9 +782,7 @@ export class WebRequestApi {
         } = details;
 
         if (WebRequestApi.isAssistantFrame(tabId, details)) {
-            logger.debug(
-                `Assistant frame detected, skipping cosmetics injection for tabId ${tabId} and frameId: ${frameId}`,
-            );
+            logger.trace(`[tsweb.WebRequestApi.injectCosmetic]: assistant frame detected, skipping cosmetics injection for tabId ${tabId} and frameId: ${frameId}`);
             return;
         }