+++ packages/tswebextension/CHANGELOG.md
@@ -41,6 +41,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [AdguardBrowserExtension#3164]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3164
 [AdguardBrowserExtension#3192]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3192
 
+### Changed
+
+- Updated [@adguard/logger] to `v2.0.0`.
+
 ## [3.1.0] - 2025-05-28
 
 ### Changed